+++ doc/changelog.rst
@@ -2,6 +2,9 @@
 What's New
 ==========
 
+0.36 (2014-06-23)
+=================
+
 * Works around ``PhantomJSRPCTimedOut`` errors.
 * Adds ``--phantomjs-exe`` option.
 * Supports extracting links from HTML ``img`` ``srcset`` attribute.